copy for all props
