+++ src/js/components/VariableMeta.js
@@ -1,11 +1,12 @@
 import React from 'react';
 import dispatcher from './../helpers/dispatcher';
 
+import CopyToClipboard from './CopyToClipboard'
 import {toType} from './../helpers/util';
 
 //icons
 import {
-    Clippy, RemoveCircle as Remove, AddCircle as Add
+    RemoveCircle as Remove, AddCircle as Add
 } from './icons';
 
 //theme
@@ -13,84 +14,6 @@ import Theme from './../themes/getStyle';
 
 
 export default class extends React.Component {
-    constructor(props) {
-        super(props);
-
-        this.state = { copied: false };
-
-        this.copiedTimer = null;
-    }
-
-    componentWillUnmount() {
-        if (this.copiedTimer) {
-            clearTimeout(this.copiedTimer);
-            this.copiedTimer = null;
-        }
-    }
-
-    handleCopy = () => {
-        const container = document.createElement('textarea');
-        const {enableClipboard, src, namespace} = this.props;
-
-        container.innerHTML = JSON.stringify(src, null, '  ');
-
-        document.body.appendChild(container);
-        container.select();
-        document.execCommand('copy');
-
-        document.body.removeChild(container);
-
-        this.copiedTimer = setTimeout(() => {
-            this.setState({
-                copied: false
-            });
-        }, 5500)
-
-        this.setState({ copied: true }, () => {
-            if (typeof enableClipboard !== "function") {
-                return;
-            }
-
-            enableClipboard({
-                src:src,
-                namespace: namespace,
-                name: namespace[namespace.length - 1]
-            });
-        })
-    }
-
-    getCopyComponent = () => {
-        const {src, size, theme, enableClipboard} = this.props;
-        let style = Theme(theme, 'copy-to-clipboard').style;
-        
-        return (
-            enableClipboard ?
-            <span class="copy-to-clipboard-container" >
-                <span
-                style={{
-                    ...style,
-                    display: 'inline-block'
-                }}
-                onClick={this.handleCopy}
-                >{this.getClippyIcon()}</span>
-            </span>
-            : null
-        );
-    }
-
-    getClippyIcon = () => {
-        const {theme} = this.props;
-
-        if (this.state.copied) {
-            return (<span>
-                <Clippy class="copy-icon" {...Theme(theme, 'copy-icon')} />
-                <span {...Theme(theme, 'copy-icon-copied')}>✔</span>
-            </span>)
-        }
-
-        return <Clippy class="copy-icon" {...Theme(theme, 'copy-icon')} />
-    }
-
     getObjectSize = () => {
         const {size, theme, displayObjectSize} = this.props;
         if (displayObjectSize) {
@@ -178,7 +101,15 @@ export default class extends React.Component {
     }
 
     render = () => {
-        const {theme, onDelete, onAdd} = this.props;
+        const {
+            theme,
+            onDelete,
+            onAdd,
+            enableClipboard,
+            src,
+            variable,
+            namespace
+        } = this.props;
         return (
         <div {...Theme(theme, 'object-meta-data')}
         class='object-meta-data'
@@ -188,7 +119,12 @@ export default class extends React.Component {
             {/* size badge display */}
             {this.getObjectSize()}
             {/* copy to clipboard icon */}
-            {this.getCopyComponent()}
+            {enableClipboard
+                ? (<CopyToClipboard
+                    clickCallback={enableClipboard}
+                    {...{src, theme, namespace}} />)
+                : null
+            }
             {/* copy add/remove icons */}
             {onAdd !== false ? this.getAddAttribute() : null}
             {onDelete !== false ? this.getRemoveObject() : null}
