+++ test/tests/js/Index-test.js
@@ -75,18 +75,18 @@ describe("<Index />", function() {
         expect(wrapper.find(".data-type-label")).to.have.length(1)
     })
 
-    it("make sure copy to clipboard is displayed for objects and arrays", function() {
+    it("make sure copy to clipboard is displayed all properties", function() {
         const wrapper = render(
             <Index
                 src={{
                     test: true,
                     passing: "hopefully",
-                    arr: [],
+                    arr: [5],
                     obj: {}
                 }}
             />
         )
-        expect(wrapper.find(".copy-to-clipboard-container")).to.have.length(3)
+        expect(wrapper.find(".copy-to-clipboard-container")).to.have.length(6)
     })
 
     it("index test componentWillReceiveProps", function() {
